Animation curve loop custom start time
Added the ability for users to set a specific time animations will loop on the animation curve, since this doesn't seem to be that easy to do with the current setup. To be clear, the animation will play from the start to the finish the first runthrough, but when reaching its end, the curve time will be reset to the specified timestamp. This property defaults to 0, maintaining the way the animation code currently works, but can also be set to -1 to automatically set the value to the end time, which means it stays on the final frame of the animation. This variable can be set for individual curves or set for the entire set.

This information is set via code or blueprints, not loaded from the GLB itself, but if anyone else has more knowledge of GLB files, they could probably set this up better.